+++ edaflow/analysis/__init__.py
@@ -4,7 +4,7 @@ Analysis module for edaflow package.
 This module contains functions for data analysis and exploration.
 """
 
-from .missing_data import (
+from .core import (
     check_null_columns,
     analyze_categorical_columns,
     convert_to_numeric,
@@ -16,7 +16,22 @@ from .missing_data import (
     handle_outliers_median,
     visualize_interactive_boxplots,
     visualize_heatmap,
-    visualize_histograms
+    visualize_histograms,
+    visualize_scatter_matrix
 )
 
-__all__ = ['check_null_columns', 'analyze_categorical_columns', 'convert_to_numeric', 'visualize_categorical_values', 'display_column_types', 'impute_numerical_median', 'impute_categorical_mode', 'visualize_numerical_boxplots', 'handle_outliers_median', 'visualize_interactive_boxplots']
+__all__ = [
+    'check_null_columns', 
+    'analyze_categorical_columns', 
+    'convert_to_numeric', 
+    'visualize_categorical_values', 
+    'display_column_types', 
+    'impute_numerical_median', 
+    'impute_categorical_mode', 
+    'visualize_numerical_boxplots', 
+    'handle_outliers_median', 
+    'visualize_interactive_boxplots',
+    'visualize_heatmap',
+    'visualize_histograms',
+    'visualize_scatter_matrix'
+]